refactor: Change project structure
With the new structure it would need an active change from the programmer to show internal classes of the package.
